better code organization
