+++ src/site/mod.rs
@@ -3,6 +3,7 @@ use crate::errors::*;
 use std::fs::File;
 use std::io::Write;
 use std::path::Path;
+pub mod artifacts;
 mod css;
 mod head;
 mod header;
